Added Frontpage label.
The label indicates whether the section's articles go to on the frontage.

Also added a .clearfix class to the grid's content paragraph and trying out small as a tag for the labels.

In hive the labels extend to the whole line height. There is no gap between the labels. Using small adds some spacing. May revert this as the small sized labels are rather unaligned.